refactor some functions into koji_habitude.cli.util
